requirements.txt
slack_bolt
python-dotenv
openai
requests
faiss-cpu
numpy